extract Season into season.rs
